fix hydration error in the official example
There is a bug in the official example. If you reload the browser window
you will see an error like "Prop `data-slate-editor-id` did not match"
at least in dev mode. It shows Slate is not working well with SSR. This
commit disables SSR for the editor.

Probably I will submit a PR to vercel/next.js to fix this after this app
is done

Signed-off-by: Vu Van Dung <me@joulev.dev>